+++ src/de/omegazirkel/risingworld/gps/exports/GpsRouteExposure.java
@@ -1,0 +1,10 @@
+package de.omegazirkel.risingworld.gps.exports;
+
+import de.omegazirkel.risingworld.gps.PluginSettings;
+
+public record GpsRouteExposure(boolean globalMarkers) {
+
+    public static GpsRouteExposure from(PluginSettings settings) {
+        return new GpsRouteExposure(settings.exposeGlobalMarkers);
+    }
+}